fix(tidy): clear pre-existing bugprone-signed-bitwise findings
Forty-five bugprone-signed-bitwise errors sat in two files that the
clang-tidy PR gate never analysed, because it only looks at files changed
against origin/master. They would have blocked the next person to edit
either file: .clang-tidy sets WarningsAsErrors: '*'.

Fixed by making the operands genuinely unsigned rather than by adding
NOLINT, matching the 0x40U convention established across the flash
executors in #294. Four sites:

- menu_actions.cpp packed-RGB unpack: mapItemColor is now unsigned, so
  the >> and & operands are unsigned; the three channel locals stay int.
- menu_actions.cpp packed-RGB compose: the two shifted channels are
  lifted into named unsigned locals.
- menu_actions.cpp haltech big-endian uint16 splits (24 lines): the
  uint16_t is cast to unsigned before shifting, and the result is cast to
  char so QByteArray::append resolves to append(char) explicitly rather
  than through an int->char conversion.
- byte_test_utils.h hex nibble join: both operands cast to unsigned.

All four are value-identical to the previous spelling. Verified two ways:
the multiset of numeric literal tokens (suffix- and case-normalised) is
unchanged in both files, and a throwaway translation unit compares the
old and new expressions exhaustively over each site's real input domain
-- the full uint16_t range for the byte splits, 0..15 squared for the
nibble join, 0..255 squared for the RGB compose, and 0x000000..0xFFFFFF
for the RGB unpack.

+++ src/algorithms/protocol/testing/byte_test_utils.h
@@ -40,7 +40,7 @@ inline bytes::Bytes bytesFromHex(std::string_view hex)
             high = nibble;
             continue;
         }
-        out.push_back(static_cast<bytes::Byte>((high << 4) | nibble));
+        out.push_back(static_cast<bytes::Byte>((static_cast<unsigned>(high) << 4U) | static_cast<unsigned>(nibble)));
         high = -1;
     }
     return out;
